test(plg): hoist esmock to before + extract shared factory for all non-admin test files (SITES-46495)
Replace per-file inline esmock maps with a shared plg-esmock-factory.js; hoist esmock
calls into before() across all remaining onboarding test files to avoid repeated module
graph instrumentation between tests.
