+++ test/controllers/plg/plg-onboarding/admin-access.test.js
@@ -13,18 +13,17 @@
 import { expect, use } from 'chai';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
-import esmock from 'esmock';
 import {
-  PLG_MODEL_DOMAIN_HELPERS,
   TEST_DOMAIN,
   TEST_IMS_ORG_ID,
-  ASO_PRODUCT_CODE,
   createSharedMocks,
+  resetStubDefaults,
   createMockSite as createMockSiteShared,
   createMockOnboarding as createMockOnboardingShared,
   createMockDataAccess,
   buildContext as buildContextShared,
 } from './shared-fixtures.js';
+import { createPlgEsmock } from './plg-esmock-factory.js';
 
 use(sinonChai);
 
@@ -33,34 +32,10 @@ describe('PlgOnboardingController', function describePlgOnboarding() {
   this.timeout(10000);
 
   let sandbox;
-
-  // Stubs for external dependencies
-  let rumRetrieveDomainkeyStub;
-  let composeBaseURLStub;
-  let detectBotBlockerStub;
-  let detectLocaleStub;
-  let resolveCanonicalUrlStub;
-  let createOrFindOrganizationStub;
-  let enableAuditsStub;
-  let enableImportsStub;
-  let triggerAuditsStub;
-  let autoResolveAuthorUrlStub;
-  let resolveWwwUrlStub;
-  let updateCodeConfigStub;
-  let findDeliveryTypeStub;
-  let deriveProjectNameStub;
-  let loadProfileConfigStub;
-  let queueDeliveryConfigWriterStub;
-  let triggerBrandProfileAgentStub;
-  let tierClientCreateForSiteStub;
-  let tierClientCreateForOrgStub;
-  let ldCreateFromStub;
-  let configToDynamoItemStub;
-  let updateRumConfigStub;
+  let stubs;
 
   // Mock objects
   let mockLog;
-  let mockEnv;
   let mockSiteConfig;
   let mockSite;
   let mockOrganization;
@@ -77,171 +52,43 @@ describe('PlgOnboardingController', function describePlgOnboarding() {
   }
 
   function buildContext(data = {}, options = {}) {
-    return buildContextShared(sandbox, mockDataAccess, mockLog, mockEnv, data, options);
+    return buildContextShared(sandbox, mockDataAccess, mockLog, stubs.mockEnv, data, options);
   }
 
-  beforeEach(async () => {
+  before(async () => {
     sandbox = sinon.createSandbox();
-
-    const shared = createSharedMocks(sandbox);
+    stubs = createSharedMocks(sandbox);
     ({
-      rumRetrieveDomainkeyStub,
-      updateRumConfigStub,
-      composeBaseURLStub,
-      detectBotBlockerStub,
-      detectLocaleStub,
-      resolveCanonicalUrlStub,
-      createOrFindOrganizationStub,
-      enableAuditsStub,
-      enableImportsStub,
-      triggerAuditsStub,
-      autoResolveAuthorUrlStub,
-      resolveWwwUrlStub,
-      updateCodeConfigStub,
-      findDeliveryTypeStub,
-      deriveProjectNameStub,
-      queueDeliveryConfigWriterStub,
-      loadProfileConfigStub,
-      triggerBrandProfileAgentStub,
-      ldCreateFromStub,
-      tierClientCreateForSiteStub,
-      tierClientCreateForOrgStub,
-      configToDynamoItemStub,
-      mockLog,
-      mockEnv,
-      mockSiteConfig,
-      mockOrganization,
-      mockProject,
-    } = shared);
+      mockSiteConfig, mockOrganization, mockProject, mockLog,
+    } = stubs);
+  });
 
-    // Default mock site (for new site flow: findByBaseURL returns null)
+  after(() => sandbox.restore());
+
+  beforeEach(() => {
+    sandbox.reset();
+    resetStubDefaults(stubs);
+    ({
+      mockSiteConfig, mockOrganization, mockProject, mockLog,
+    } = stubs);
+
     mockSite = createMockSite();
-
-    // PlgOnboarding mock
     mockOnboarding = createMockOnboarding();
-
-    // DataAccess
     mockDataAccess = createMockDataAccess(sandbox, {
       mockSite, mockOrganization, mockProject, mockOnboarding,
     });
   });
 
-  afterEach(() => {
-    sandbox.restore();
-  });
-
   describe('onboard - admin access', () => {
     let adminController;
+    let AdminPlgOnboardingControllerFactory;
 
-    beforeEach(async () => {
-      const AdminPlgOnboardingController = (await esmock(
-        '../../../../src/controllers/plg/plg-onboarding.js',
-        {
-          '@adobe/spacecat-shared-utils': {
-            composeBaseURL: composeBaseURLStub,
-            detectBotBlocker: detectBotBlockerStub,
-            detectLocale: detectLocaleStub,
-            hasText: (val) => typeof val === 'string' && val.trim().length > 0,
-            isValidIMSOrgId: (val) => typeof val === 'string' && val.endsWith('@AdobeOrg'),
-            resolveCanonicalUrl: resolveCanonicalUrlStub,
-          },
-          '@adobe/spacecat-shared-http-utils': {
-            badRequest: (msg) => ({ status: 400, value: msg }),
-            createResponse: (body, status) => ({ status, value: body }),
-            forbidden: (msg) => ({ status: 403, value: msg }),
-            internalServerError: (msg) => ({ status: 500, value: msg }),
-            notFound: (msg) => ({ status: 404, value: msg }),
-            ok: (data) => ({ status: 200, value: data }),
-          },
-          '@adobe/spacecat-shared-rum-api-client': {
-            default: {
-              createFrom: sandbox.stub().returns({
-                retrieveDomainkey: rumRetrieveDomainkeyStub,
-              }),
-            },
-          },
-          '@adobe/spacecat-shared-launchdarkly-client': {
-            default: ldCreateFromStub,
-          },
-          '@adobe/spacecat-shared-tier-client': {
-            default: {
-              createForSite: tierClientCreateForSiteStub,
-              createForOrg: tierClientCreateForOrgStub,
-            },
-          },
-          '@adobe/spacecat-shared-data-access/src/models/site/config.js': {
-            Config: { toDynamoItem: configToDynamoItemStub },
-          },
-          '@adobe/spacecat-shared-data-access/src/models/entitlement/index.js': {
-            Entitlement: {
-              PRODUCT_CODES: { ASO: ASO_PRODUCT_CODE },
-              TIERS: {
-                FREE_TRIAL: 'FREE_TRIAL', PAID: 'PAID', PLG: 'PLG', PRE_ONBOARD: 'PRE_ONBOARD',
-              },
-            },
-          },
-          '@adobe/spacecat-shared-data-access/src/models/plg-onboarding/plg-onboarding.model.js': {
-            default: {
-              ...PLG_MODEL_DOMAIN_HELPERS,
-              STATUSES: {
-                IN_PROGRESS: 'IN_PROGRESS',
-                ONBOARDED: 'ONBOARDED',
-                PRE_ONBOARDING: 'PRE_ONBOARDING',
-                ERROR: 'ERROR',
-                WAITING_FOR_IP_ALLOWLISTING: 'WAITING_FOR_IP_ALLOWLISTING',
-                WAITLISTED: 'WAITLISTED',
-                INACTIVE: 'INACTIVE',
-                REJECTED: 'REJECTED',
-                OUTDATED: 'OUTDATED',
-              },
-              REVIEW_REASONS: {
-                DOMAIN_ALREADY_ONBOARDED_IN_ORG: 'DOMAIN_ALREADY_ONBOARDED_IN_ORG',
-                AEM_SITE_CHECK: 'AEM_SITE_CHECK',
-                DOMAIN_ALREADY_ASSIGNED: 'DOMAIN_ALREADY_ASSIGNED',
-                BOT_BLOCKER: 'BOT_BLOCKER',
-              },
-              REVIEW_DECISIONS: {
-                BYPASSED: 'BYPASSED',
-                UPHELD: 'UPHELD',
-                CLOSED: 'CLOSED',
-                REOPENED: 'REOPENED',
-                OFFBOARDED: 'OFFBOARDED',
-                PENDING: 'PENDING',
-              },
-            },
-          },
-          '../../../../src/controllers/llmo/llmo-onboarding.js': {
-            createOrFindOrganization: createOrFindOrganizationStub,
-            enableAudits: enableAuditsStub,
-            enableImports: enableImportsStub,
-            triggerAudits: triggerAuditsStub,
-          },
-          '../../../../src/support/utils.js': {
-            autoResolveAuthorUrl: autoResolveAuthorUrlStub,
-            resolveWwwUrl: resolveWwwUrlStub,
-            updateCodeConfig: updateCodeConfigStub,
-            findDeliveryType: findDeliveryTypeStub,
-            deriveProjectName: deriveProjectNameStub,
-            queueDeliveryConfigWriter: queueDeliveryConfigWriterStub,
-          },
-          '../../../../src/utils/slack/base.js': {
-            loadProfileConfig: loadProfileConfigStub,
-          },
-          '../../../../src/support/brand-profile-trigger.js': {
-            triggerBrandProfileAgent: triggerBrandProfileAgentStub,
-          },
-          '../../../../src/support/access-control-util.js': {
-            default: {
-              fromContext: () => ({ hasAdminAccess: () => true }),
-            },
-          },
-          '../../../../src/support/rum-config-service.js': {
-            updateRumConfig: updateRumConfigStub,
-          },
-        },
-      )).default;
+    before(async () => {
+      AdminPlgOnboardingControllerFactory = await createPlgEsmock(stubs, { hasAdminAccess: true });
+    });
 
-      adminController = AdminPlgOnboardingController({ log: mockLog });
+    beforeEach(() => {
+      adminController = AdminPlgOnboardingControllerFactory({ log: mockLog });
     });
 
     it('returns 400 when imsOrgId is missing in admin onboard call', async () => {
@@ -262,7 +109,7 @@ describe('PlgOnboardingController', function describePlgOnboarding() {
       const context = buildContext({ domain: TEST_DOMAIN, imsOrgId: TEST_IMS_ORG_ID });
       const res = await adminController.onboard(context);
       expect(res.status).to.equal(200);
-      expect(createOrFindOrganizationStub).to.have.been.calledWith(
+      expect(stubs.createOrFindOrganizationStub).to.have.been.calledWith(
         TEST_IMS_ORG_ID,
         sinon.match.any,
       );
