+++ test/controllers/plg/plg-onboarding/admin-get-status.test.js
@@ -23,7 +23,7 @@ import {
   createMockDataAccess,
   mockAuthInfo as mockAuthInfoShared,
 } from './shared-fixtures.js';
-import { createPlgEsmock } from './admin-esmock-factory.js';
+import { createPlgEsmock } from './plg-esmock-factory.js';
 
 use(sinonChai);
 
@@ -81,6 +81,7 @@ describe('PlgOnboardingController - getStatus', function () {
     });
     stubs.ldGetFeatureFlagStub.resolves({ variations: [{ value: {} }] });
     stubs.rumRetrieveDomainkeyStub.resolves('test-domainkey');
+    stubs.rumApiClientCreateFromStub.returns({ retrieveDomainkey: stubs.rumRetrieveDomainkeyStub });
     stubs.detectBotBlockerStub.resolves({ crawlable: true });
     stubs.findDeliveryTypeStub.resolves('aem_edge');
     stubs.deriveProjectNameStub.returns('example.com');
