+++ test/controllers/plg/plg-onboarding/admin-update.test.js
@@ -28,7 +28,7 @@ import {
   createMockOnboarding as createMockOnboardingShared,
   createMockDataAccess,
 } from './shared-fixtures.js';
-import { createPlgEsmock } from './admin-esmock-factory.js';
+import { createPlgEsmock } from './plg-esmock-factory.js';
 
 use(sinonChai);
 
@@ -82,6 +82,7 @@ describe('PlgOnboardingController - update', function () {
       updateVariationValue: stubs.ldUpdateVariationValueStub,
     });
     stubs.rumRetrieveDomainkeyStub.resolves('test-domainkey');
+    stubs.rumApiClientCreateFromStub.returns({ retrieveDomainkey: stubs.rumRetrieveDomainkeyStub });
     stubs.detectBotBlockerStub.resolves({ crawlable: true });
     stubs.findDeliveryTypeStub.resolves('aem_edge');
     stubs.deriveProjectNameStub.returns('example.com');
